MM-124 refactor: 개인 일정 조회 URI 변경 (#139)
