+++ src/main/java/com/dev/moim/domain/moim/service/CalenderQueryService.java
@@ -7,8 +7,6 @@ import java.util.List;
 
 public interface CalenderQueryService {
 
-    PlanMonthListDTO<List<UserPlanDTO>> getIndividualPlans(User user, int year, int month);
-
     PlanMonthListDTO<List<UserPlanDTO>> getUserPlans(User user, int year, int month);
 
     PlanMonthListDTO<PlanDayListDTO> getMoimPlans(User user, Long moimId, int year, int month);
