+++ src/main/java/com/dev/moim/domain/user/service/UserQueryService.java
@@ -1,7 +1,9 @@
 package com.dev.moim.domain.user.service;
 
 import com.dev.moim.domain.account.entity.User;
+import com.dev.moim.domain.moim.dto.calender.PlanMonthListDTO;
 import com.dev.moim.domain.moim.dto.calender.UserDailyPlanPageDTO;
+import com.dev.moim.domain.moim.dto.calender.UserPlanDTO;
 import com.dev.moim.domain.user.dto.ChatRoomUserListResponse;
 import com.dev.moim.domain.user.dto.ProfileDTO;
 import com.dev.moim.domain.user.dto.ProfileDetailDTO;
@@ -20,6 +22,8 @@ public interface UserQueryService {
     UserDailyPlanPageDTO getUserDailyMoimPlan(User user, int year, int month, int day, int page, int size);
 
     UserDailyPlanPageDTO getUserDailyIndividualPlan(User user, int year, int month, int day, int page, int size);
+
+    PlanMonthListDTO<List<UserPlanDTO>> getIndividualPlans(User user, int year, int month);
 
     List<Long> findUserMoimIdListByUserId(Long userId);
